subject line
Redo calibration for new setup

That was done mainly through the function recalibrate in Quantification.py, plus some small acompanying changes in Molecules and Object_files. I think I also made lots of small changes in plotting and selecting functions, such as select_cycle cutting MS data properly for the first time. All of this is being done in connection with sniff2paper, and the recent installation of new chip new interface block new fillament etc. The changes in this commit are quick and dirty for now with a lot in the front-panel scripts, but it'll need to be systematized soon. I'm planning for my data treatment methods here to set industry standards and revolutionize my field of electrocatalysis. Still to do from today:
Formalize the M18 correction to calibrations.
Write a Chip class
Interface with PyExpLabSys and cinfdata
Make EC_MS more transparent to users. Clearer code.

Scott out.