Make Chips module and do external calibration
Chips class added with a function for calculating the capillary flux according to Daniel Trimarco's PhD thesis. This allows for external calibrations, as implemented in the Calibrations module. All molecule data files were moved to their own subfolder and rewritten using new functions in the Molecules module. A quadrupole transmission function can be entered in the Molecule.get_RSF() to better predict sensitivities. The recalibrate function in the Quantification module was expanded (it has become quite a mess) to make figure 5b of our Trimarco and Scott et al 2017, and used to generate the new values of F_cal in all of the Molecule data files. Some standard colors changed.