Cinfdata. Epoch times. Triggers. Shunt correction.
Okay, huge overhaul of EC_MS, major improvements to the functionality of the package. There's now an importing function, download_cinfdata_set, which interfaces with cinfdata.py so that I will never have to copy text from cinfdata ever again. All of the major upgrades to timestamp handling (epoch time) and improvements in functionality, organization, and readability of the synchronize function are now copied over from EC_Xray. In addition to the time_cal function from EC_Xray, there is now a trigger_cal function which more or less automatically lines up the potentiostat and MS data using the TTL triggers sent from the potentiostat to the Analog In channel of the QMS data. Finally, there are functions for calculating and correcting for a shunt current pathway between CE and ground, which I proved in the middle of the night 18D05 to be the mechanism of "tiltyness" or "slantyness" of our CV's. These functions are all fully vectorized, and I even improved the vectorization of some others including select_cycles. There might be some small patches to old functions to make them work smoothely with all these major improvements. I'm starting to think about making classes to handle the data rather than continuing to rely on these somewhat clunky python dictionaries for everything in this package. But that's far in the future. It'll also be interesting to see what Spectro comes up with. Scott out.