function save_as_text implemented
--- What need was addressed ---
The need to save processed data.
--- What was done ----
function save_as_text in module Data_Importing was written to export data from an EC_MS dataset as text in a versitile whay.
--- What still needs to be done
This, like so many others, should be a function of a DataSet class, rather than just clumsily using dictionaries for datasets as if they were a custom class.