Union tspan when datasets are appended
When appending datasets with Combining.synchronize(), tspan[-1] was less than tspan[0] in the combined dataset because it tried to calculate the intersection. Now the function calculates a union tspan, give by [t_begin, t_end], which is used as the default tspan when append=True.

Git also seems to recognize a lot of lines in Combining.py and EC.py as changed which appear un-changed. Maybe it's due to spyder now removing trailing whitespaces.